cambios

+++ Taller3_2do2021/src/dominio/Documento.java
@@ -25,10 +25,10 @@ public class Documento extends Entrega{
 	}
 	@Override
 	public int pagar() {
-		int precioPeso = peso/1000;
-		int precioGrosor = grosor/10;
-		int precioPagar = precioPeso*precioGrosor*100;
-		return precioPagar;
+		double precioPeso = peso/1000.0;
+		double precioGrosor = grosor/10.0;
+		double precioPagar = precioPeso*precioGrosor*100;
+		return (int)precioPagar;
 	}	
 			
 }
